Move agent to first target.

+++ meaoo-mqtt/src/main/kotlin/fr/nodesigner/meaoo/mqtt/android/model/MqttModel.kt
@@ -1,17 +1,17 @@
 package fr.nodesigner.meaoo.mqtt.android.model
 
 data class Path(
-        val vehicle_type: String?,
-        val target: Coordinate?
+    val vehicle_type: String?,
+    val target: Coordinate?
 )
 
 data class Coordinate(
-        val x: Float,
-        val y: Float
+    val x: Double,
+    val y: Double
 )
 
 data class Teleport(
-        val vehicle_type: String?,
-        val costs: List<Float>,
-        val path: List<List<Float>>
+    val vehicle_type: String?,
+    val costs: List<Double>,
+    val path: List<List<Double>>
 )
